django/requirements.txt
Django==4.2.2
djangorestframework
markdowncd
django-filter
djoser
django-cors-headers
psycopg2-binary
pillow
stripe